Group the optimizer components into a package
Summary:
@tianyu-l asked for a dedicated folder for the optimizer components, and both
@fegin and @tianyu-l pointed out that `checkpoint_utils.py` is not an optimizer
module even though three of its four public functions are optimizer-only.

Split it on that line and group the result:

  components/optimizer.py                 -> components/optimizer/optimizer.py
  components/lr_scheduler.py              -> components/optimizer/lr_scheduler.py
  components/checkpoint_utils.py:optimizer state helpers
                                          -> components/optimizer/utils.py

`init_optim_state`, `get_flat_optim_state_dict`, and `load_flat_optim_state_dict`
all take a `torch.optim.Optimizer`, so they move. `canonical_fqn(name: str, ...)`
does not: it strips activation-checkpoint wrapper segments from a dotted FQN and
is a checkpointing concern. It stays in `components/checkpoint_utils.py` here and
moves to `components/checkpointer/utils.py` in the next change, which is the one
that creates that folder, so it only ever moves once.

Nothing outside the package has to change its imports:

- `components/optimizer/__init__.py` re-exports `OptimizersContainer`,
  `ParamGroupConfig`, `default_adamw`, `register_moe_load_balancing_hook`, and
  `LRSchedulersContainer`, so `torchtitan.components.optimizer` keeps working and
  the scheduler now comes from the same short path as the container.
- `components/lr_scheduler.py` and `components/checkpoint_utils.py` stay as
  forwarding modules for their old names.

EMA has no module today, so there is nothing to move for it yet.

Test Plan:
`pytest tests/unit_tests/test_checkpoint.py tests/unit_tests/test_state_dict_keys.py
tests/unit_tests/test_lr_scheduler.py tests/unit_tests/test_optimizer_param_groups.py`
passes (68 tests) at this commit, in a venv matching the CPU CI job.

`torchtitan.trainer` and `torchtitan.components.optimizer` both import standalone
at this commit. `flake8` and `ufmt` (pinned black 22.12.0 / usort 1.0.5, matching
`.pre-commit-config.yaml`) are clean on the changed files.
