+++ torchtitan/components/optimizer/__init__.py
@@ -1,0 +1,21 @@
+# Copyright (c) Meta Platforms, Inc. and affiliates.
+# All rights reserved.
+#
+# This source code is licensed under the BSD-style license found in the
+# LICENSE file in the root directory of this source tree.
+
+from .lr_scheduler import LRSchedulersContainer
+from .optimizer import (
+    default_adamw,
+    OptimizersContainer,
+    ParamGroupConfig,
+    register_moe_load_balancing_hook,
+)
+
+__all__ = [
+    "LRSchedulersContainer",
+    "OptimizersContainer",
+    "ParamGroupConfig",
+    "default_adamw",
+    "register_moe_load_balancing_hook",
+]
